requirements.txt
google-generativeai
streamlit
python-dotenv
pandas
plotly